requirements.txt
# requirements.txt - LLM Code Assistant 依存関係リスト
# 本番環境用の固定バージョン

# Core Framework & GUI
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# LLM & AI Libraries
openai==1.12.0
anthropic==0.18.1
transformers==4.38.2
torch==2.2.1
tokenizers==0.15.2
accelerate==0.27.2
sentence-transformers==2.5.1
llama-cpp-python==0.2.55
ctransformers==0.2.27

# Vector Database & Embeddings
faiss-cpu==1.7.4
chromadb==0.4.24
numpy==1.26.4
scipy==1.12.0

# Database & Storage
SQLAlchemy==2.0.28
alembic==1.13.1

# Web Framework & API
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.4
httpx==0.27.0
aiohttp==3.9.3
requests==2.31.0

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.1
toml==0.10.2

# File Processing & Text
nltk==3.9.1
spacy==3.8.7
python-magic==0.4.27
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
markdown==3.6
beautifulsoup4==4.12.3

# Code Analysis
autopep8==2.0.4
black==24.2.0
flake8==7.0.0
pylint==3.1.0
mypy==1.8.0

# Syntax Highlighting & Code Editor
Pygments==2.17.2
QScintilla==2.14.1

# Git Integration
GitPython==3.1.42

# Encryption & Security
cryptography==42.0.5
bcrypt==4.1.2
keyring==24.3.1

# Logging & Monitoring
loguru==0.7.2
structlog==24.1.0
psutil==6.1.0

# Async & Concurrency
aiofiles==23.2.1

# Image Processing
Pillow==10.2.0

# Template Engine
Jinja2==3.1.3

# Plugin System
pluggy==1.4.0

# Utilities
tqdm==4.66.2
colorama==0.4.6
click==8.1.7
rich==13.7.1
python-dateutil==2.9.0
watchdog==4.0.0
typing-extensions==4.10.0
